[Chore] #57 MyReview rating 타입 Double로 변경

+++ DaOnGil/app/src/main/java/kr/tekit/lion/daongil/presentation/myreview/adapter/MyReviewRVAdapter.kt
@@ -42,7 +42,7 @@ class MyReviewRVAdapter(private val myReviewList: List<MyReview>, private val it
             binding.textViewMyReviewLocation.text = myReview.myReviewLocation
             binding.textViewMyReviewLocationName.text = myReview.myReviewLocationName
 
-            settingRating(myReview.rating.toDouble())
+            settingRating(myReview.rating)
         }
 
         private fun settingRating(rating: Double) {
